refactor: simplify IM platforms to DingTalk/QQ and make paths platform-aware
- Remove Telegram, Discord, Slack from IM bridge (keep DingTalk and QQ)
- Add platform-specific home directory: LOCALAPPDATA on Windows, ~/.enikk on Unix
- Update README and config examples to reflect IM platform changes
- Fix mypy type hints in ui_parser.py

+++ enikk/ui_parser.py
@@ -13,7 +13,7 @@ logger = logging.getLogger(__name__)
 
 
 def _letterbox(img: np.ndarray, new_shape: tuple[int, int] = (640, 640),
-               color: tuple[int, int, int] = (114, 114, 114)) -> tuple[np.ndarray, float, tuple[int, int]]:
+               color: tuple[int, int, int] = (114, 114, 114)) -> tuple[np.ndarray, float, tuple[float, float]]:
     """Resize image with padding to maintain aspect ratio.
 
     Returns:
